styles and contact form
