+++ package.json
@@ -29,10 +29,10 @@
     "gatsby-transformer-json": "^2.11.0",
     "gatsby-transformer-remark": "^2.16.0",
     "gatsby-transformer-sharp": "^2.5.2",
-    "netlify-cms": "^2.10.90",
+    "netlify-cms": "^2.10.93",
     "prop-types": "^15.7.2",
     "react": "^16.12.0",
-    "react-alice-carousel": "^2.3.0",
+    "react-alice-carousel": "^2.4.0",
     "react-dom": "^16.12.0",
     "react-google-recaptcha": "^2.1.0",
     "react-helmet": "^6.0.0",
